Modify project environment so examples can use CPU
This only affects the dev container. It consists of further
commands run in the script launched by the onCreateCommand.

The core package does not depend on torch (PyTorch), but the
examples do. By default, on GNU/Linux systems, torch expects to use
CUDA. This installs the CPU build of torch, replacing the one
poetry picks (because the GNU/Linux CPU build is not from PyPI).

Modifying pyproject.toml or poetry.lock to achieve this would get
in the way of using this dev container for development. To avoid
that, I use the somewhat ugly technique of running pip in the
poetry-managed environment.

It is possible to use a GPU in a dev container, and it's possible
someone is using this project together with work that benefits from
one. But this project does not require a GPU, and most use of dev
containers is in codespaces that don't offer a GPU by default and
where a codespace supporting GPU computations is more expensive.
So having PyTorch use the CPU to do computations is a reasonable
default, for a dev container whose chief purpose is to allow people
to get up and running fast.