Use 3.10 in dev container to support old code too
This changes the dev container's Debian 12 based Python image from
the Python 3.11 image to the Python 3.10 image, so that people who
want to use original_codebase/ code can do so in the dev container,
without having to install a different Python implementation.

This is only to make the dev container *compatible* with the old
codebase as well as the new. Unlike the new codebase, the original
codebase is not being *installed* automatically during container
creation, because doing so would increase the startup time of the
dev container. (It would also increase the size of the associated
storage volume during use, which is relevant to cost in some uses
of codespaces.)