Do dynamic settings update in postCreateCommand
Updating defaultInterpreterPath in the remote settings file
dynamically in the script run by the onCreateCommand didn't work
because the file was overwritten based on the (other) settings
specified statically in devcontainer.json and various defaults.

Normally this problem would be avoided by including the desired
value as a static setting in devcontainer.json. But the correct
defaultInterpreterPath cannot be specified statically in a
non-brittle way, because it contains the poetry-produced hash,
which depends on paths that may change, and which uses an algorithm
that may not be guaranteed to remain the same across versions of
poetry (though it rarely changes).